requirements.txt
Django>=5.0,<6.0
psycopg[binary]>=3.1
requests>=2.31
python-dotenv>=1.0
dj-database-url>=2.2
djangorestframework>=3.15
django-auditlog>=3.0
Pillow>=10.0
reportlab>=4.2
celery>=5.4
redis>=5.0
